+++ src/entities/driver.hpp
@@ -18,7 +18,7 @@ public:
         int age,
         int work_experience,
         int children_count,
-        float salary,
+        double salary,
         std::vector<int>& passed_medexam
     );
